pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "Shiparr"
version = "0.1.0"
description = "Self-hosted GitOps deployment service for Docker Compose, designed for the *arr ecosystem."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
  { name = "Your Name", email = "you@example.com" },
]

# Runtime dependencies (section 7 - Dépendances Python)
dependencies = [
  "quart>=0.19",
  "quart-auth>=0.9",
  "sqlalchemy[asyncio]>=2.0",
  "aiosqlite>=0.19",
  "apscheduler>=3.10",
  "gitpython>=3.1",
  "docker>=7.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "pydantic>=2.5",
  "pydantic-settings>=2.1",
  "structlog>=24.1",
]

[project.optional-dependencies]
# Development dependencies (section 7 - dev dependencies)
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "ruff>=0.3",
]

[project.urls]
Homepage = "https://example.com/Shiparr"
Repository = "https://example.com/Shiparr/repo"

[tool.hatch.build.targets.wheel]
packages = ["src/Shiparr"]

[tool.pytest.ini_options]
addopts = "-v --cov=Shiparr --cov-report=term-missing"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B"]

[tool.coverage.run]
source = ["Shiparr"]
branch = true
